refactor : supabase related functions
